player search p1
I want to add player search because there is a possibility of 100-200 players showing on the page and scrolling sucks I want to make it so that people can search it. I addded a Player class to the index.js becuase that will deal with all the players stat because of the rankings. this will also make it easier for me to add css things because I am not rendering text anymore